Move inits to Sentry files
